# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delilah_resto",
    "description": "Food Restaurant API",
    "contact": { "email": "jta1223@gmail.com" },
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/user/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "List User",
        "operationId": "listUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/listUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Updated User",
        "operationId": "udpatedUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updatedUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/listUser"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["User"],
        "summary": "Deleted User",
        "operationId": "deletedUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "User Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/order/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Order"],
        "summary": "List order",
        "description": "List order in our app",
        "operationId": "listOrder",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succesfull operation",
            "schema": {
              "$ref": "#/definitions/listOrder"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Order"],
        "summary": "Deleted order",
        "description": "Deleted order in our app",
        "operationId": "deletedOrder",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "type": "integer",
            "name": "id",
            "description": "Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succesfull operation"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter your bearer token in the format JWT"
    }
  },
  "definitions": {
...
    "listUser": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "Jonathan Jimenez S"
          },
          "fullname": {
            "type": "string",
            "example": "jta1223"
          },
          "email": {
            "type": "string",
            "example": "admin@admin.com"
          },
          "phone": {
            "type": "string",
            "example": "1234567890"
          },
          "address": {
            "type": "string",
            "example": "Cl 10 50 90"
          },
          "admin": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "enable": {
            "format": "int32",
            "type": "integer",
            "example": 1
          },
          "createAt": {
            "type": "string",
            "example": "2020-12-05T02:08:39.000Z"
          }
        }
      }
    },
...
    "updatedUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "Jonathan Jimenez S"
        },
        "password": {
          "type": "string",
          "example": "admin123"
        },
        "fullname": {
          "type": "string",
          "example": "jta1223"
        },
        "email": {
          "type": "string",
          "example": "admin@admin.com"
        },
        "phone": {
          "type": "string",
          "example": "1234567890"
        },
        "address": {
          "type": "string",
          "example": "Cl 10 50 90"
        }
      }
    },
...
    "listOrder": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "created": {
            "type": "string",
            "example": "2020-12-05T21:19:38.000Z"
          },
          "User": {
            "type": "object",
            "properties": {
              "fullname": {
                "type": "string",
                "example": "Jonathan Jimenez S"
              },
              "username": {
                "type": "string",
                "example": "jta1223"
              },
              "email": {
                "type": "string",
                "example": "jta1223@gmail.com"
              },
              "phone": {
                "type": "string",
                "example": "1234567890"
              },
              "address": {
                "type": "string",
                "example": "Cl 10 50 90"
              }
            }
          },
          "Status": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "NUEVO"
              }
            }
          },
          "Payment": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "EFECTIVO"
              }
            }
          },
          "Order_Products": {
            "type": "array",
            "items": {
              "properties": {
                "quantity": {
                  "type": "integer",
                  "format": "int32",
                  "example": 1
                },
                "price": {
                  "type": "integer",
                  "format": "int32",
                  "example": 1
                },
                "Product": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Sandwich de Focaccia"
                    },
                    "img": {
                      "type": "string",
                      "example": "https://www.petitchef.es/imgupl/recipe/focaccia-sandwich--md-166713p249402.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}